Adding player validity checks
Addons like Wiremod call the standard GM hooks like CanTool and
GravGunPunt when their Wire Forcers and Grabbers try to push/pull
entities.

This is fine until the player that owns the Wire entity disconnects from
the server. The GM hooks get a NULL player entity, which makes it
impossible to check if the owner had access. The only way I know that
could solve this is if Wiremod used their own hooks and passed the Wire
entity along with it, allowing Aegis to get the owner from it.

For now, we'll just block these sorts of events. There isn't much else
we can do on our end.